+++ growthmate-api/src/main/java/com/growup/growthmate/query/api/CommunityQueryV1Controller.java
@@ -3,6 +3,11 @@ package com.growup.growthmate.query.api;
 import com.growup.growthmate.LoginMember;
 import com.growup.growthmate.query.application.CommunityQueryService;
 import com.growup.growthmate.query.dto.*;
+import com.growup.growthmate.query.dto.request.CommentQueryRequest;
+import com.growup.growthmate.query.dto.request.PostDetailRequest;
+import com.growup.growthmate.query.dto.response.CommentResponse;
+import com.growup.growthmate.query.dto.response.PostDetailResponse;
+import com.growup.growthmate.query.dto.response.PostPreviewResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
