+++ growthmate-api/src/main/java/com/growup/growthmate/query/application/CommunityQueryService.java
@@ -3,6 +3,11 @@ package com.growup.growthmate.query.application;
 import com.growup.growthmate.BusinessException;
 import com.growup.growthmate.community.post.exception.PostException;
 import com.growup.growthmate.query.dto.*;
+import com.growup.growthmate.query.dto.request.CommentQueryRequest;
+import com.growup.growthmate.query.dto.request.PostDetailRequest;
+import com.growup.growthmate.query.dto.response.CommentResponse;
+import com.growup.growthmate.query.dto.response.PostDetailResponse;
+import com.growup.growthmate.query.dto.response.PostPreviewResponse;
 import com.growup.growthmate.query.repository.projection.CommentProjection;
 import com.growup.growthmate.query.repository.projection.PostDetailProjection;
 import com.growup.growthmate.query.repository.CommunityQueryRepository;
