+++ growthmate-api/src/main/java/com/growup/growthmate/query/dto/request/CommentQueryRequest.java
@@ -1,4 +1,4 @@
-package com.growup.growthmate.query.dto;
+package com.growup.growthmate.query.dto.request;
 
 public record CommentQueryRequest(Long postId, Long loginId) {
 }
