+++ growthmate-api/src/test/java/com/growup/growthmate/query/application/CommunityQueryServiceTest.java
@@ -4,9 +4,9 @@ import com.growup.growthmate.BusinessException;
 import com.growup.growthmate.community.post.exception.PostException;
 import com.growup.growthmate.isolation.TestIsolation;
 import com.growup.growthmate.query.dto.PagingParams;
-import com.growup.growthmate.query.dto.PostDetailRequest;
-import com.growup.growthmate.query.dto.PostDetailResponse;
-import com.growup.growthmate.query.dto.PostPreviewResponse;
+import com.growup.growthmate.query.dto.request.PostDetailRequest;
+import com.growup.growthmate.query.dto.response.PostDetailResponse;
+import com.growup.growthmate.query.dto.response.PostPreviewResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
